modified files have been uploaded for the second iteration
changed the method of recognizing objects, removing excess in the image. Added calculation of the areas of objects and polygons. 
TODO: the phone's camera was not very good, so some items in the photo are not recognized well, so I will update all the photos in the future. Added angle recognition, but it doesn't work very well yet.